requirements.txt
openai==1.12.0
anthropic==0.8.1
requests==2.31.0
moviepy==1.0.3
Pillow==10.0.0
python-dotenv==1.0.0
pyperclip==1.8.2
pydub==0.25.1
gTTS==2.3.2
elevenlabs==1.13.5
azure-cognitiveservices-speech==1.32.1

# Note: FFmpeg is required for audio processing
# Install FFmpeg from https://ffmpeg.org/download.html
# For Windows users: choco install ffmpeg 